Treats strings with underscores as singular words
The problem with the previous approach is the vim's colorscheme inability to parse such words properly. During development, if a string contains multiple keywords separated with underscores, for example in cpp return_true_path, it would be weirdly highlighted and add a lot of unnecessary mess to the code. It's kinda handy, to be able to jump in string between substrings separated by underscores, but the cost of this is too much for me.