Uses global variables to set various variables specific to local setup
Common configuration is stored in the .vimrc dotfile. Different machines can store this file in different places, so to accomodate to that, the example script presents which variables should be set to point to different relevant directories, like the location of Python or UltiSnips's snippets. I also didn't want to use environment variables to facilitate that, as I think that the usage of Vim variables provides a more encapsulated approach.