Unifies background coloring
Makes sure that the background is always the same on the line numbers and the text itself. Unfortunatelly that required also some tweaks with Goyo. The screen flashes frantically when switching between normal mode and 'distract free' mode, but I think that the amount of flashing lights during the toggle makes it the 'distract amplify' mode, lol.