Silences functions that fetch color information
For some reason, interaction with DLLs and plugins is messy. Or maybe it's just changing colors and window properties at the same time? I don't know, but sometimes an output of ':highlight ColorColumn' or something else was being displayed and it even required confirmation... Current solution silences that output and removes the need for confirmation, which makes the experience of getting into and out of Goyo way smoother.