Fixes toggle error when trying to toggle freshly exited console
As with all the code taken directly from the internet, one may get into trouble very quickly. In this case, after closing Windows terminal with 'exit' command, I tried to toggle the console and instead of console, I saw an error related to buffer which wasn't found, but was wanted for deletion. This commit fixes that problem.