Enables GoToReference functionality in YCM
This commit doesn't do it per se, but it enables it. After searching for a while, and sifting through YCM documentation / Gitter, I noticed that people you clangd instead of clang as a backend and also -background-index is passed to it to enable GoToReference. I don't know whether it's true, but you know what? It works and that's all I need :)