Introduce function the columnizes initialization
By columnization, I mean having variables aligned in a column fashion.
This is a first implementation, so it only has one mode - the
columnization pivot stays in place, but the variables are aligned in
columnic (?) fashion. It would be nice to also have other modes, like
aligning the end of variables instead of beginning.